+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## 2.0.7
+- Added support for Android 12 OS (BroadCast receiver)
+
+## 2.0.6
+- Added support for Android 12 OS
+
+## 2.0.5
+- Fixed issue with callback on 'scannerReady' event
+
+## 2.0.4
+- Added callback for prominent disclosure message, to notify Flutter app
+
+## 2.0.3
+- Updated flow for showing the prominent disclosure message to be only shown when invoked from method channel
+
+## 2.0.1
+- Added method to add custom beacons layout (Android)
+- Added method to set foreground scan periods (Android)
+- Added method to set background scan periods (Android)
+- Updated example app, added list of monitored beacons
+
 ## 2.0.0
 - Migrated to null safety